Move Button from start to gameboard
